Update vc-shortcode-autoloader.php
Removed additional if statement to line 67 as unrequired.
The shorthand for isset does what the original code intended. i.e. if self::$config['classmap'][ $class ] is the same as isset(self::$config['classmap'][ $class ]). If it does not exist then isset will fail, as will isset fail if it does exist but is null.